+++ src/headers/src/ProjectInfoPopup.cpp
@@ -4,6 +4,7 @@
 
 #include "../../incl/Avalanche.hpp"
 
+#include <iomanip>
 #include <sstream>
 
 #include <fmt/core.h>
@@ -21,23 +22,30 @@
 using namespace geode::prelude;
 using namespace avalanche;
 
-ProjectInfoPopup* ProjectInfoPopup::create() {
-  auto ret = new ProjectInfoPopup;
-  if (ret->initAnchored(440, 290)) {
-    ret->autorelease();
-    return ret;
+Handler* avalHandler = Handler::get();
+
+inline std::string url_encode(const std::string& value) {
+  std::ostringstream escaped;
+  escaped.fill('0');
+  escaped << std::hex;
+
+  for (char c : value) {
+    if (isalnum(static_cast<unsigned char>(c)) || c == '-' || c == '_' || c == '.' || c == '~') {
+      escaped << c;
+    } else {
+      escaped << '%' << std::setw(2) << int((unsigned char)c);
+    };
   };
 
-  CC_SAFE_DELETE(ret);
-  return nullptr;
+  return escaped.str();
 };
 
-void ProjectInfoPopup::infoPopup(CCObject*) {
+void ProjectInfoPopup::doInfo(Project proj, std::string publisher) {
   std::ostringstream typeOfProj;
 
-  switch (m_avalProject.type) {
+  switch (proj.type) {
   case Project::Type::TEAM:
-    typeOfProj << "a <cg>team project</c> hosted by <cy>" << m_avalProject.host << "</c>";
+    typeOfProj << "a <cg>team project</c> hosted by <cy>" << proj.host << "</c>";
     break;
 
   case Project::Type::COLLAB:
@@ -58,27 +66,74 @@ void ProjectInfoPopup::infoPopup(CCObject*) {
   };
 
   std::ostringstream body;
-  body << "<cy>" << m_avalPublisher << "</c> - '<cg>" << m_avalProject.name << "</c>' is " << typeOfProj.str() << ". You can watch its showcase here.";
+  body << "<cy>" << publisher << "</c> - '<cg>" << proj.name << "</c>' is " << typeOfProj.str() << ". You can watch its showcase here.";
 
   std::string resultBody = body.str();
 
   createQuickPopup(
-    m_avalProject.name.c_str(),
+    proj.name.c_str(),
     resultBody.c_str(),
     "OK", "Watch",
-    [this](auto, bool btn2) {
+    [proj](auto, bool btn2) {
       if (btn2) {
-        AVAL_LOG_INFO("Opening showcase link in browser: {}", this->m_avalProject.showcase_url);
-        web::openLinkInBrowser(this->m_avalProject.showcase_url);
+        AVAL_LOG_INFO("Opening showcase link in browser: {}", proj.showcase);
+        web::openLinkInBrowser(proj.showcase);
       } else {
         AVAL_LOG_DEBUG("User clicked OK");
       }; },
     true);
 };
 
+void ProjectInfoPopup::doShowcase(Project proj, std::string publisher) {
+  std::ostringstream body;
+  body << "Watch the full showcase video for <cy>" << publisher << "</c> - '<cg>" << proj.name << "</c>'?";
+
+  std::string resultBody = body.str();
+
+  createQuickPopup(
+    proj.name.c_str(),
+    resultBody.c_str(),
+    "Cancel", "Watch",
+    [proj](auto, bool btn2) {
+      if (btn2) {
+        AVAL_LOG_INFO("Opening showcase link in browser: {}", proj.showcase);
+        web::openLinkInBrowser(proj.showcase);
+      } else {
+        AVAL_LOG_DEBUG("User clicked Cancel");
+      }; },
+    true);
+};
+
+ProjectInfoPopup* ProjectInfoPopup::create() {
+  auto ret = new ProjectInfoPopup;
+  if (ret->initAnchored(440, 290)) {
+    ret->autorelease();
+    return ret;
+  };
+
+  CC_SAFE_DELETE(ret);
+  return nullptr;
+};
+
+void ProjectInfoPopup::infoPopup(CCObject*) {
+  ProjectInfoPopup::doInfo(m_avalProject, m_avalPublisher);
+};
+
+void ProjectInfoPopup::onPlayShowcase(CCObject*) {
+  ProjectInfoPopup::doShowcase(m_avalProject, m_avalPublisher);
+};
+
+void ProjectInfoPopup::infoPopupLinked(CCObject*) {
+  ProjectInfoPopup::doInfo(m_linkedProject, m_linkedPublisher);
+};
+
+void ProjectInfoPopup::onPlayShowcaseLinked(CCObject*) {
+  ProjectInfoPopup::doShowcase(m_linkedProject, m_linkedPublisher);
+};
+
 void ProjectInfoPopup::onFameInfo(CCObject*) {
   std::ostringstream body;
-  body << "This level, '<cg>" << m_avalProject.name << "</c>', is featured in <cl>Avalanche's</c> <cy>Hall of Fame</c>. It is a special list of levels that are considered to be the best of the best from the team.";
+  body << "The level '<cg>" << m_avalProject.name << "</c>' by <cy>" << m_avalPublisher << "</c> is featured in <cl>Avalanche's</c> <cy>Hall of Fame</c>. It is a special list of levels that are considered to be the best of the best from the team.";
 
   std::string resultBody = body.str();
 
@@ -92,26 +147,6 @@ void ProjectInfoPopup::onFameInfo(CCObject*) {
         web::openLinkInBrowser(URL_AVALANCHE);
       } else {
         AVAL_LOG_DEBUG("User clicked OK");
-      }; },
-    true);
-};
-
-void ProjectInfoPopup::onPlayShowcase(CCObject*) {
-  std::ostringstream body;
-  body << "Watch the full showcase video for <cy>" << m_avalPublisher << "</c> - '<cg>" << m_avalProject.name << "</c>'?";
-
-  std::string resultBody = body.str();
-
-  createQuickPopup(
-    m_avalProject.name.c_str(),
-    resultBody.c_str(),
-    "Cancel", "Watch",
-    [this](auto, bool btn2) {
-      if (btn2) {
-        AVAL_LOG_INFO("Opening showcase link in browser: {}", this->m_avalProject.showcase_url);
-        web::openLinkInBrowser(this->m_avalProject.showcase_url);
-      } else {
-        AVAL_LOG_DEBUG("User clicked Cancel");
       }; },
     true);
 };
@@ -164,7 +199,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   setZOrder(10);
 
   m_level = level;
-  m_avalProject = Handler::get()->GetProject(m_level->m_levelID.value());
+  m_avalProject = avalHandler->GetProject(m_level->m_levelID.value());
 
   if (m_avalProject.type == Project::Type::NONE) {
     AVAL_LOG_ERROR("Avalanche project type is NONE");
@@ -210,6 +245,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
     fameIcon->setScale(0.25f);
 
     auto fameLabel = CCLabelBMFont::create("Avalanche Hall of Fame", "goldFont.fnt");
+    fameLabel->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
     fameLabel->ignoreAnchorPointForPosition(false);
     fameLabel->setAnchorPoint({ 0.5, 0.5 });
     fameLabel->setScale(0.375f);
@@ -265,7 +301,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   art_bottomLeft->setID("bottom-left-corner");
   art_bottomLeft->setAnchorPoint({ 0, 0 });
   art_bottomLeft->setPosition({ 0, 0 });
-  art_bottomLeft->setScale(1.250);
+  art_bottomLeft->setScale(1.250f);
   art_bottomLeft->setFlipX(false);
   art_bottomLeft->setFlipY(false);
   art_bottomLeft->setZOrder(0);
@@ -276,7 +312,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   art_bottomRight->setID("bottom-right-corner");
   art_bottomRight->setAnchorPoint({ 1, 0 });
   art_bottomRight->setPosition({ m_overlayMenu->getScaledContentWidth(), 0 });
-  art_bottomRight->setScale(1.250);
+  art_bottomRight->setScale(1.250f);
   art_bottomRight->setFlipX(true);
   art_bottomRight->setFlipY(false);
   art_bottomLeft->setZOrder(0);
@@ -287,7 +323,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   art_topLeft->setID("top-left-corner");
   art_topLeft->setAnchorPoint({ 0, 1 });
   art_topLeft->setPosition({ 0, m_overlayMenu->getScaledContentHeight() });
-  art_topLeft->setScale(1.250);
+  art_topLeft->setScale(1.250f);
   art_topLeft->setFlipX(false);
   art_topLeft->setFlipY(true);
   art_topLeft->setZOrder(0);
@@ -298,7 +334,7 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
   art_topRight->setID("top-right-corner");
   art_topRight->setAnchorPoint({ 1, 1 });
   art_topRight->setPosition({ m_overlayMenu->getScaledContentWidth(), m_overlayMenu->getScaledContentHeight() });
-  art_topRight->setScale(1.250);
+  art_topRight->setScale(1.250f);
   art_topRight->setFlipX(true);
   art_topRight->setFlipY(true);
   art_topRight->setZOrder(0);
@@ -331,27 +367,163 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
 
   auto hostName_label = CCLabelBMFont::create(hostLabelTxt, "bigFont.fnt");
   hostName_label->setID("host-name-label");
+  hostName_label->setAnchorPoint({ 0, 1 });
   hostName_label->ignoreAnchorPointForPosition(false);
-  hostName_label->setAnchorPoint({ 0, 0.5 });
-  hostName_label->setPosition({ 10.f, (m_mainLayer->getScaledContentHeight() / 2.f) + 70.f });
+  hostName_label->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
+  hostName_label->setPosition({ 10.f, m_mainLayer->getScaledContentHeight() - 55.f });
   hostName_label->setScale(0.25f);
 
   auto hostName = CCLabelBMFont::create(m_avalProject.host.c_str(), "goldFont.fnt");
   hostName->setID("host-name");
+  hostName->setAnchorPoint({ 0, 1 });
   hostName->ignoreAnchorPointForPosition(false);
-  hostName->setAnchorPoint({ 0, 0.5 });
-  hostName->setPosition({ 10.f, (m_mainLayer->getScaledContentHeight() / 2.f) + 55.f });
+  hostName->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
+  hostName->setPosition({ 10.f, m_mainLayer->getScaledContentHeight() - 65.f });
   hostName->setScale(0.75f);
 
   m_overlayMenu->addChild(hostName_label);
   m_overlayMenu->addChild(hostName);
 
-  auto playShowcase_label = CCLabelBMFont::create("Watch the Showcase", "chatFont.fnt");
+  AVAL_LOG_INFO("Adding showcase container to main project '{}'", m_avalProject.name);
+
+  // create showcase project button
+  AVAL_LOG_DEBUG("Creating showcase project button");
+  auto showcaseProjMenu = CCMenu::create();
+  showcaseProjMenu->setID("showcase-project-menu");
+  showcaseProjMenu->setAnchorPoint({ 0.5, 0 });
+  showcaseProjMenu->ignoreAnchorPointForPosition(false);
+  showcaseProjMenu->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, 12.5f });
+  showcaseProjMenu->setScaledContentSize({ 192.f, 108.f });
+  showcaseProjMenu->setZOrder(2);
+
+  m_clippingNode->addChild(showcaseProjMenu);
+
+  // create clipping node for showcase project
+  AVAL_LOG_DEBUG("Creating clipping node for showcase project");
+  auto showcaseProjClippingNode = CCClippingNode::create();
+  showcaseProjClippingNode->setID("clipping-node");
+  showcaseProjClippingNode->setContentSize(showcaseProjMenu->getScaledContentSize());
+  showcaseProjClippingNode->ignoreAnchorPointForPosition(false);
+  showcaseProjClippingNode->setAnchorPoint({ 0.5f, 0.5f });
+  showcaseProjClippingNode->setPosition({ showcaseProjMenu->getScaledContentWidth() / 2.f, showcaseProjMenu->getScaledContentHeight() / 2.f });
+  showcaseProjClippingNode->setStencil(CCLayerColor::create({ 250, 250, 250 }, showcaseProjMenu->getScaledContentWidth(), showcaseProjMenu->getScaledContentHeight()));;
+  showcaseProjClippingNode->setZOrder(-1);
+
+  showcaseProjMenu->addChild(showcaseProjClippingNode);
+
+  auto showcaseProjClippingNodeBg = CCScale9Sprite::create("GJ_square04.png");
+  showcaseProjClippingNodeBg->setID("background");
+  showcaseProjClippingNodeBg->setContentSize(showcaseProjClippingNode->getContentSize());
+  showcaseProjClippingNodeBg->setPosition(showcaseProjClippingNode->getPosition());
+  showcaseProjClippingNodeBg->setAnchorPoint(showcaseProjClippingNode->getAnchorPoint());
+  showcaseProjClippingNodeBg->setZOrder(-1);
+
+  showcaseProjClippingNode->addChild(showcaseProjClippingNodeBg);
+
+  // corner art deco for showcase project
+  auto corner = "rewardCorner_001.png";
+
+  auto art_bottomLeft_showcaseProj = CCSprite::createWithSpriteFrameName(corner);
+  art_bottomLeft_showcaseProj->setID("bottom-left-corner");
+  art_bottomLeft_showcaseProj->setAnchorPoint({ 0, 0 });
+  art_bottomLeft_showcaseProj->setPosition({ 0, 0 });
+  art_bottomLeft_showcaseProj->setScale(0.75f);
+  art_bottomLeft_showcaseProj->setFlipX(false);
+  art_bottomLeft_showcaseProj->setFlipY(false);
+  art_bottomLeft_showcaseProj->setZOrder(4);
+
+  showcaseProjClippingNode->addChild(art_bottomLeft_showcaseProj);
+
+  auto art_bottomRight_showcaseProj = CCSprite::createWithSpriteFrameName(corner);
+  art_bottomRight_showcaseProj->setID("bottom-right-corner");
+  art_bottomRight_showcaseProj->setAnchorPoint({ 1, 0 });
+  art_bottomRight_showcaseProj->setPosition({ showcaseProjClippingNode->getScaledContentWidth(), 0 });
+  art_bottomRight_showcaseProj->setScale(0.75f);
+  art_bottomRight_showcaseProj->setFlipX(true);
+  art_bottomRight_showcaseProj->setFlipY(false);
+  art_bottomRight_showcaseProj->setZOrder(4);
+
+  showcaseProjClippingNode->addChild(art_bottomRight_showcaseProj);
+
+  auto art_topLeft_showcaseProj = CCSprite::createWithSpriteFrameName(corner);
+  art_topLeft_showcaseProj->setID("top-left-corner");
+  art_topLeft_showcaseProj->setAnchorPoint({ 0, 1 });
+  art_topLeft_showcaseProj->setPosition({ 0, showcaseProjClippingNode->getScaledContentHeight() });
+  art_topLeft_showcaseProj->setScale(0.75f);
+  art_topLeft_showcaseProj->setFlipX(false);
+  art_topLeft_showcaseProj->setFlipY(true);
+  art_topLeft_showcaseProj->setZOrder(4);
+
+  showcaseProjClippingNode->addChild(art_topLeft_showcaseProj);
+
+  auto art_topRight_showcaseProj = CCSprite::createWithSpriteFrameName(corner);
+  art_topRight_showcaseProj->setID("top-right-corner");
+  art_topRight_showcaseProj->setAnchorPoint({ 1, 1 });
+  art_topRight_showcaseProj->setPosition({ showcaseProjClippingNode->getScaledContentWidth(), showcaseProjClippingNode->getScaledContentHeight() });
+  art_topRight_showcaseProj->setScale(0.75f);
+  art_topRight_showcaseProj->setFlipX(true);
+  art_topRight_showcaseProj->setFlipY(true);
+  art_topRight_showcaseProj->setZOrder(4);
+
+  showcaseProjClippingNode->addChild(art_topRight_showcaseProj);
+
+  // create thumbnail lazy sprite for showcase project
+  if (AVAL_GEODE_MOD->getSettingValue<bool>("show-proj-thumb")) {
+    AVAL_LOG_DEBUG("Creating thumbnail lazy sprite for showcase project '{}'", m_avalProject.name);
+    LazySprite* showcaseProjThumb = LazySprite::create(showcaseProjMenu->getScaledContentSize(), true);
+    showcaseProjThumb->setID("thumbnail");
+    showcaseProjThumb->setAnchorPoint({ 0.5, 0.5 });
+    showcaseProjThumb->ignoreAnchorPointForPosition(false);
+    showcaseProjThumb->setPosition({ showcaseProjMenu->getScaledContentWidth() / 2.f, showcaseProjMenu->getScaledContentHeight() / 2.f });
+    showcaseProjThumb->setScale(0.5f);
+
+    showcaseProjThumb->setLoadCallback([showcaseProjThumb, showcaseProjClippingNode](Result<> res) {
+      if (res.isOk()) {
+        AVAL_LOG_INFO("Showcase project thumbnail loaded successfully");
+
+        showcaseProjThumb->setScale(1.f);
+        showcaseProjThumb->setScale(showcaseProjClippingNode->getScaledContentHeight() / showcaseProjThumb->getScaledContentHeight());
+
+        showcaseProjThumb->setPosition(showcaseProjClippingNode->getPosition());
+        showcaseProjThumb->ignoreAnchorPointForPosition(false);
+        showcaseProjThumb->setColor({ 125, 125, 125 });
+        showcaseProjThumb->setOpacity(125);
+      } else {
+        AVAL_LOG_ERROR("Failed to load showcase project thumbnail: {}", res.unwrapErr());
+        showcaseProjThumb->removeMeAndCleanup();
+      };
+                                       });
+
+    std::string encodedShowcaseUrl = url_encode(m_avalProject.showcase); // encode the showcase url for use in the thumbnail url
+    std::string showcaseProjThumbURL = fmt::format("https://api.cubicstudios.xyz/avalanche/v1/fetch/yt-thumbnails?url={}", (std::string)encodedShowcaseUrl); // custom thumbnail
+
+    AVAL_LOG_DEBUG("Getting showcase project thumbnail at {}...", (std::string)showcaseProjThumbURL);
+    showcaseProjThumb->loadFromUrl(showcaseProjThumbURL, LazySprite::Format::kFmtUnKnown, false);
+    if (showcaseProjThumb) showcaseProjClippingNode->addChild(showcaseProjThumb);
+  } else {
+    AVAL_LOG_DEBUG("Showcase project thumbnail setting is disabled, not adding thumbnail to showcase containered project container");
+  };
+
+  // set border
+  auto showcaseProjBorder = CCScale9Sprite::create("GJ_square07.png");
+  showcaseProjBorder->setID("border");
+  showcaseProjBorder->setPosition(showcaseProjClippingNode->getPosition());
+  showcaseProjBorder->setContentSize(showcaseProjClippingNode->getScaledContentSize());
+  showcaseProjBorder->ignoreAnchorPointForPosition(false);
+  showcaseProjBorder->setAnchorPoint({ 0.5f, 0.5f });
+  showcaseProjBorder->setZOrder(3);
+
+  // add border to clipping node
+  showcaseProjClippingNode->addChild(showcaseProjBorder);
+
+  auto playShowcase_label = CCLabelBMFont::create("Watch the\nShowcase", "bigFont.fnt");
   playShowcase_label->setID("play-showcase-label");
-  playShowcase_label->ignoreAnchorPointForPosition(false);
+  playShowcase_label->setScale(0.5f);
   playShowcase_label->setAnchorPoint({ 0.5, 0.5 });
-  playShowcase_label->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, 60.f });
-  playShowcase_label->setScale(1.f);
+  playShowcase_label->ignoreAnchorPointForPosition(false);
+  playShowcase_label->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
+  playShowcase_label->setPosition({ showcaseProjMenu->getScaledContentWidth() / 2.f, 75.f });
+  playShowcase_label->setZOrder(3);
 
   auto playShowcase_sprite = CCSprite::createWithSpriteFrameName("GJ_playBtn2_001.png");
   playShowcase_sprite->setScale(0.5f);
@@ -360,11 +532,12 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
     playShowcase_sprite,
     this,
     menu_selector(ProjectInfoPopup::onPlayShowcase));
-  playShowcase->setID("play-showcase-button");
-  playShowcase->setPosition({ m_mainLayer->getScaledContentWidth() / 2.f, 30.f });
+  playShowcase->setID("play-showcase");
+  playShowcase->setPosition({ showcaseProjMenu->getScaledContentWidth() / 2.f, 37.5f });
+  playShowcase->setZOrder(4);
 
-  m_overlayMenu->addChild(playShowcase_label);
-  m_overlayMenu->addChild(playShowcase);
+  showcaseProjMenu->addChild(playShowcase_label);
+  showcaseProjMenu->addChild(playShowcase);
 
   // project thumbnail
   if (AVAL_GEODE_MOD->getSettingValue<bool>("show-proj-thumb")) {
@@ -377,18 +550,6 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
     projThumb->ignoreAnchorPointForPosition(false);
     projThumb->setPosition({ m_clippingNode->getScaledContentWidth() / 2, m_clippingNode->getScaledContentHeight() / 2 });
 
-    // extract video id from url
-    std::string videoId = m_avalProject.showcase_url;
-
-    std::string prefix = "https://youtu.be/";
-
-    if (videoId.find(prefix) == 0) videoId = videoId.substr(prefix.length());
-
-    // format url
-    auto projThumbURL = fmt::format("https://img.youtube.com/vi/{}/maxresdefault.jpg", (std::string)videoId);
-
-    AVAL_LOG_DEBUG("Getting thumbnail at {}...", projThumbURL);
-
     projThumb->setLoadCallback([this, projThumb, bgSize](Result<> res) {
       if (res.isOk()) {
         AVAL_LOG_INFO("Sprite loaded successfully");
@@ -397,21 +558,214 @@ ProjectInfoPopup* ProjectInfoPopup::setProject(GJGameLevel* level) {
         float scale = bgSize.height / projThumb->getContentHeight();
 
         projThumb->setScale(scale);
-        projThumb->setAnchorPoint({ 0, 0 });
         projThumb->ignoreAnchorPointForPosition(false);
         projThumb->setColor({ 125, 125, 125 });
-        projThumb->setPosition({ 0, 0 });
         projThumb->setOpacity(125);
       } else {
-        AVAL_LOG_ERROR("{}", res.unwrapErr());
+        AVAL_LOG_ERROR("Failed to get project thumbnail, {}", res.unwrapErr());
         projThumb->removeMeAndCleanup();
       };
                                });
 
+    std::string projThumbURL = fmt::format("https://api.cubicstudios.xyz/avalanche/v1/fetch/thumbnails?id={}", (int)m_level->m_levelID.value()); // custom thumbnail
+
+    AVAL_LOG_DEBUG("Getting thumbnail at {}...", (std::string)projThumbURL);
     projThumb->loadFromUrl(projThumbURL, LazySprite::Format::kFmtUnKnown, false);
     if (projThumb) m_clippingNode->addChild(projThumb);
   } else {
     AVAL_LOG_DEBUG("Project thumbnail setting is disabled, not adding thumbnail to project info popup");
+  };
+
+  if (m_avalProject.link_to_main.enabled) {
+    AVAL_LOG_DEBUG("Project '{}' has a link to the main project", m_avalProject.name);
+    m_linkedProject = avalHandler->GetProject(m_avalProject.link_to_main.level_id);
+
+    if (m_linkedProject.type == Project::Type::NONE) {
+      AVAL_LOG_ERROR("Failed to get linked project with ID {}", m_avalProject.link_to_main.level_id);
+    } else {
+      AVAL_LOG_INFO("Adding link to main project '{}'", m_linkedProject.name);
+
+      // create linked project button
+      AVAL_LOG_DEBUG("Creating linked project button");
+      auto linkedProjMenu = CCMenu::create();
+      linkedProjMenu->setID("linked-project-menu");
+      linkedProjMenu->setAnchorPoint({ 1, 1 });
+      linkedProjMenu->ignoreAnchorPointForPosition(false);
+      linkedProjMenu->setPosition({ m_mainLayer->getScaledContentWidth() - 25.f, m_mainLayer->getScaledContentHeight() - 55.f });
+      linkedProjMenu->setScaledContentSize({ 150.f, 81.f });
+      linkedProjMenu->setZOrder(2);
+
+      m_clippingNode->addChild(linkedProjMenu);
+
+      // info button
+      auto linkedProjInfoBtnSprite = CCSprite::createWithSpriteFrameName("GJ_infoIcon_001.png");
+      linkedProjInfoBtnSprite->setScale(0.375f);
+
+      auto linkedProjInfoBtn = CCMenuItemSpriteExtra::create(
+        linkedProjInfoBtnSprite,
+        this,
+        menu_selector(ProjectInfoPopup::infoPopupLinked));
+      linkedProjInfoBtn->setID("info-button");
+      linkedProjInfoBtn->setPosition({ linkedProjMenu->getScaledContentWidth() - 6.25f, linkedProjMenu->getScaledContentHeight() - 6.25f });
+      linkedProjInfoBtn->setZOrder(126);
+
+      linkedProjMenu->addChild(linkedProjInfoBtn);
+
+      // create clipping node for linked project
+      AVAL_LOG_DEBUG("Creating clipping node for linked project");
+      auto linkedProjClippingNode = CCClippingNode::create();
+      linkedProjClippingNode->setID("clipping-node");
+      linkedProjClippingNode->setContentSize(linkedProjMenu->getScaledContentSize());
+      linkedProjClippingNode->ignoreAnchorPointForPosition(false);
+      linkedProjClippingNode->setAnchorPoint({ 0.5f, 0.5f });
+      linkedProjClippingNode->setPosition({ linkedProjMenu->getScaledContentWidth() / 2.f, linkedProjMenu->getScaledContentHeight() / 2.f });
+      linkedProjClippingNode->setStencil(CCLayerColor::create({ 250, 250, 250 }, linkedProjMenu->getScaledContentWidth(), linkedProjMenu->getScaledContentHeight()));;
+      linkedProjClippingNode->setZOrder(-1);
+
+      linkedProjMenu->addChild(linkedProjClippingNode);
+
+      auto linkedProjClippingNodeBg = CCScale9Sprite::create("GJ_square02.png");
+      linkedProjClippingNodeBg->setID("background");
+      linkedProjClippingNodeBg->setContentSize(linkedProjClippingNode->getContentSize());
+      linkedProjClippingNodeBg->setPosition(linkedProjClippingNode->getPosition());
+      linkedProjClippingNodeBg->setAnchorPoint(linkedProjClippingNode->getAnchorPoint());
+      linkedProjClippingNodeBg->setZOrder(-1);
+
+      linkedProjClippingNode->addChild(linkedProjClippingNodeBg);
+
+      // corner art deco for linked project
+      auto corner = "rewardCorner_001.png";
+
+      auto art_bottomLeft_linkedProj = CCSprite::createWithSpriteFrameName(corner);
+      art_bottomLeft_linkedProj->setID("bottom-left-corner");
+      art_bottomLeft_linkedProj->setAnchorPoint({ 0, 0 });
+      art_bottomLeft_linkedProj->setPosition({ 0, 0 });
+      art_bottomLeft_linkedProj->setScale(0.5f);
+      art_bottomLeft_linkedProj->setFlipX(false);
+      art_bottomLeft_linkedProj->setFlipY(false);
+      art_bottomLeft_linkedProj->setZOrder(3);
+
+      linkedProjClippingNode->addChild(art_bottomLeft_linkedProj);
+
+      auto art_bottomRight_linkedProj = CCSprite::createWithSpriteFrameName(corner);
+      art_bottomRight_linkedProj->setID("bottom-right-corner");
+      art_bottomRight_linkedProj->setAnchorPoint({ 1, 0 });
+      art_bottomRight_linkedProj->setPosition({ linkedProjClippingNode->getScaledContentWidth(), 0 });
+      art_bottomRight_linkedProj->setScale(0.5f);
+      art_bottomRight_linkedProj->setFlipX(true);
+      art_bottomRight_linkedProj->setFlipY(false);
+      art_bottomRight_linkedProj->setZOrder(3);
+
+      linkedProjClippingNode->addChild(art_bottomRight_linkedProj);
+
+      auto art_topLeft_linkedProj = CCSprite::createWithSpriteFrameName(corner);
+      art_topLeft_linkedProj->setID("top-left-corner");
+      art_topLeft_linkedProj->setAnchorPoint({ 0, 1 });
+      art_topLeft_linkedProj->setPosition({ 0, linkedProjClippingNode->getScaledContentHeight() });
+      art_topLeft_linkedProj->setScale(0.5f);
+      art_topLeft_linkedProj->setFlipX(false);
+      art_topLeft_linkedProj->setFlipY(true);
+      art_topLeft_linkedProj->setZOrder(3);
+
+      linkedProjClippingNode->addChild(art_topLeft_linkedProj);
+
+      auto art_topRight_linkedProj = CCSprite::createWithSpriteFrameName(corner);
+      art_topRight_linkedProj->setID("top-right-corner");
+      art_topRight_linkedProj->setAnchorPoint({ 1, 1 });
+      art_topRight_linkedProj->setPosition({ linkedProjClippingNode->getScaledContentWidth(), linkedProjClippingNode->getScaledContentHeight() });
+      art_topRight_linkedProj->setScale(0.5f);
+      art_topRight_linkedProj->setFlipX(true);
+      art_topRight_linkedProj->setFlipY(true);
+      art_topRight_linkedProj->setZOrder(3);
+
+      linkedProjClippingNode->addChild(art_topRight_linkedProj);
+
+      // create thumbnail lazy sprite for linked project
+      if (AVAL_GEODE_MOD->getSettingValue<bool>("show-proj-thumb")) {
+        AVAL_LOG_DEBUG("Creating thumbnail lazy sprite for linked project '{}'", m_linkedProject.name);
+        LazySprite* linkedProjThumb = LazySprite::create(linkedProjMenu->getScaledContentSize(), true);
+        linkedProjThumb->setID("thumbnail");
+        linkedProjThumb->setAnchorPoint({ 0.5, 0.5 });
+        linkedProjThumb->ignoreAnchorPointForPosition(false);
+        linkedProjThumb->setPosition({ linkedProjMenu->getScaledContentWidth() / 2.f, linkedProjMenu->getScaledContentHeight() / 2.f });
+        linkedProjThumb->setScale(0.5f);
+
+        linkedProjThumb->setLoadCallback([linkedProjThumb, linkedProjClippingNode](Result<> res) {
+          if (res.isOk()) {
+            AVAL_LOG_INFO("Linked project thumbnail loaded successfully");
+
+            linkedProjThumb->setScale(1.f);
+            linkedProjThumb->setScale(linkedProjClippingNode->getScaledContentHeight() / linkedProjThumb->getScaledContentHeight());
+
+            linkedProjThumb->setPosition(linkedProjClippingNode->getPosition());
+            linkedProjThumb->ignoreAnchorPointForPosition(false);
+            linkedProjThumb->setColor({ 250, 250, 250 });
+            linkedProjThumb->setOpacity(250);
+          } else {
+            AVAL_LOG_ERROR("Failed to load linked project thumbnail: {}", res.unwrapErr());
+            linkedProjThumb->removeMeAndCleanup();
+          };
+                                         });
+
+        std::string encodedShowcaseUrl = url_encode(m_linkedProject.showcase); // encode the showcase url for use in the thumbnail url
+        std::string linkedProjThumbURL = fmt::format("https://api.cubicstudios.xyz/avalanche/v1/fetch/yt-thumbnails?url={}", (std::string)encodedShowcaseUrl); // custom thumbnail
+
+        AVAL_LOG_DEBUG("Getting linked project thumbnail at {}...", (std::string)linkedProjThumbURL);
+        linkedProjThumb->loadFromUrl(linkedProjThumbURL, LazySprite::Format::kFmtUnKnown, false);
+        if (linkedProjThumb) linkedProjClippingNode->addChild(linkedProjThumb);
+      } else {
+        AVAL_LOG_DEBUG("Linked project thumbnail setting is disabled, not adding thumbnail to linked project container");
+      };
+
+      // set border
+      auto linkedProjBorder = CCScale9Sprite::create("GJ_square07.png");
+      linkedProjBorder->setID("border");
+      linkedProjBorder->setPosition(linkedProjClippingNode->getPosition());
+      linkedProjBorder->setContentSize(linkedProjClippingNode->getScaledContentSize());
+      linkedProjBorder->ignoreAnchorPointForPosition(false);
+      linkedProjBorder->setAnchorPoint({ 0.5f, 0.5f });
+      linkedProjBorder->setZOrder(2);
+
+      // add border to clipping node
+      linkedProjClippingNode->addChild(linkedProjBorder);
+
+      // create sprite for linked project showcase button
+      auto linkedProjShowcase_sprite = CCSprite::createWithSpriteFrameName("GJ_playBtn2_001.png");
+      linkedProjShowcase_sprite->setScale(0.375f);
+
+      // create button to play linked project showcase
+      AVAL_LOG_DEBUG("Creating button to play linked project showcase");
+      auto linkedProjShowcase = CCMenuItemSpriteExtra::create(
+        linkedProjShowcase_sprite,
+        this,
+        menu_selector(ProjectInfoPopup::onPlayShowcaseLinked));
+      linkedProjShowcase->setID("button");
+      linkedProjShowcase->setPosition({ linkedProjMenu->getScaledContentWidth() / 2.f, 25.f });
+
+      linkedProjMenu->addChild(linkedProjShowcase);
+
+      // create label for linked project showcase
+      AVAL_LOG_DEBUG("Creating label for linked project showcase");
+      auto linkedProjLabel = CCLabelBMFont::create("Play Now!", "bigFont.fnt");
+      linkedProjLabel->setID("label");
+      linkedProjLabel->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
+      linkedProjLabel->setPosition({ linkedProjMenu->getScaledContentWidth() / 2.f, linkedProjMenu->getScaledContentHeight() - 10.f });
+      linkedProjLabel->setScale(0.25f);
+
+      linkedProjMenu->addChild(linkedProjLabel);
+
+      // create name text label for linked project name
+      AVAL_LOG_DEBUG("Creating label for linked project name");
+      auto linkedProjName = CCLabelBMFont::create(m_linkedProject.name.c_str(), "goldFont.fnt");
+      linkedProjName->setID("name");
+      linkedProjName->setAlignment(CCTextAlignment::kCCTextAlignmentCenter);
+      linkedProjName->setPosition({ linkedProjMenu->getScaledContentWidth() / 2.f, linkedProjMenu->getScaledContentHeight() - 25.f });
+      linkedProjName->setScale(0.625f);
+
+      linkedProjMenu->addChild(linkedProjName);
+    };
+  } else {
+    AVAL_LOG_DEBUG("Project '{}' does not have a link to the main project", m_avalProject.name);
   };
 
   // geode settings popup button