+++ changelog.md
@@ -1,3 +1,19 @@
+## v1.2.0
+#### Team Project pop-up
+
+#### Changes
+- **Updated** Geode compatibility to version `4.6.3`
+- **Added** level thumbnail background to Team Project information pop-up
+- **Added** promotional container for parent Team Project, if any, in information pop-up for independent levels
+- **Reworked** level showcase menu in Team Project information pop-up
+
+#### Developers
+- **Updated** URLs to use Cubic Studios API endpoints for all web requests
+
+###### Latest
+---
+###### Older
+
 ### v1.1.2
 ##### Patches
 
@@ -6,10 +22,6 @@
 - **Added** toggle setting `Collaboration Level Cells`
 - **Added** toggle setting `Collaboration Level Pages`
 - Minor tweaks
-
-###### Latest
----
-###### Older
 
 ### v1.1.1
 ##### Patches